backend/requirements.txt
# --- Web Framework ---
fastapi==0.136.0
uvicorn==0.44.0
python-multipart==0.0.26
pydantic==2.13.3

# --- Database & Auth ---
supabase==2.28.3

# --- PDF Generation ---
weasyprint==68.1
Jinja2==3.1.6

# --- AI Extractor ---
google-generativeai==0.8.6
anthropic==0.96.0

# --- Excel & Google Drive ---
openpyxl==3.1.5
google-api-python-client==2.194.0
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1

# --- Utilities ---
python-dotenv==1.2.2